Move converters to their own modules
Why: in line with moving everything to wrappers for uniformity, the
converters have been changed.
