+++ src/orion/core/worker/consumer.py
@@ -14,7 +14,7 @@ import subprocess
 import sys
 import tempfile
 
-from orion.core.io.convert import JSONConverter
+from orion.core.io.converters.jsonconverter import JSONConverter
 from orion.core.io.database import Database
 from orion.core.io.space_builder import SpaceBuilder
 from orion.core.worker.trial import Trial
